Add files via upload
This dataset contains comprehensive information related to advertising campaigns conducted by our marketing agency. It includes features such as user demographics, ad content, time of display, and the binary indicator of whether the user clicked on the advertisement or not. The dataset serves as the foundation for developing a machine learning algorithm aimed at predicting user engagement and optimising ad click-through rates. Carefully curated and preprocessed, this data provides valuable insights for training and evaluating predictive models, contributing to the enhancement of advertising campaign strategies.